fix: mobx + refactor some things
