+++ apps/frontend/src/pages/games/galaki/GalakiPage.tsx
@@ -1,67 +1,15 @@
 import BackIcon from "@mui/icons-material/ChevronLeft";
 import { Stack, Typography, useTheme } from "@mui/material";
+import { observer } from "mobx-react-lite";
+import { type FC, useMemo } from "react";
+import { RenouvArtWait } from "../../../reusableComponents/animations/RenouvArtWait/RenouvArtWait";
 import { CenteredDiv } from "../../../reusableComponents/common/CenteredDiv";
 import {
     GaladrimButton,
     type GaladrimButtonProps,
 } from "../../../reusableComponents/common/GaladrimButton";
 import { RoundedLinks } from "../../../reusableComponents/common/RoundedLinks";
-import { type FC, useMemo } from "react";
-import { APPLICATION_JSON_HEADERS } from "../../idea/createIdea/CreateIdeaStore";
-import { fetchBackendJson } from "../../../api/fetch";
-import { makeAutoObservable } from "mobx";
-import { _assert } from "@galadrim-tools/shared";
-import { RenouvArtWait } from "../../../reusableComponents/animations/RenouvArtWait/RenouvArtWait";
-import { observer } from "mobx-react-lite";
-
-type Guess = {
-    id: number;
-    portraitGuessable: {
-        pictureUrl: string;
-        guess: string;
-    };
-};
-
-class GalakiStore {
-    guesses: Guess[] | undefined;
-    currentGuessIdx = 0;
-    constructor() {
-        makeAutoObservable(this);
-    }
-
-    async fetchGuesses() {
-        const fetchResult = await fetchBackendJson("/portraitGuessGame", "GET", {
-            headers: APPLICATION_JSON_HEADERS,
-        });
-        if (fetchResult.ok) {
-            this.guesses = (await fetchResult.json) as Guess[];
-        }
-    }
-
-    get fetchingGuess() {
-        return this.guesses === undefined;
-    }
-
-    async setGuess(grade: 1 | 2 | 3 | 4) {
-        _assert(this.currentGuess);
-        const body = JSON.stringify({
-            guessId: this.currentGuess.id,
-            grade,
-        });
-        fetchBackendJson("/portraitGuessGame", "POST", {
-            body,
-            headers: APPLICATION_JSON_HEADERS,
-        });
-    }
-
-    get currentGuess() {
-        return this.guesses?.at(this.currentGuessIdx);
-    }
-
-    processNextGuess() {
-        this.currentGuessIdx += 1;
-    }
-}
+import { GalakiStore } from "./GalakiStore";
 
 const GalakiPage = () => {
     const guessStore = useMemo(() => {
